requirements.txt
openai-whisper
gradio
gTTS